adjust initial menu
